requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.9.4
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
asteroid-filterbanks==0.4.0
astroid==2.15.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
autopage==0.5.1
backcall==0.2.0
backports.cached-property==1.0.2
beautifulsoup4==4.12.0
bitarray==2.7.3
blessed==1.20.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
cliff==4.2.0
cmaes==0.9.1
cmake==3.26.1
cmd2==2.4.3
colorama==0.4.6
colorlog==6.7.0
comm==0.1.3
commonmark==0.9.1
contourpy==1.0.7
croniter==1.3.8
cycler==0.11.0
Cython==0.29.33
cytoolz==0.12.1
dataclasses==0.6
dateutils==0.6.12
debugpy==1.6.6
decorator==5.1.1
deepdiff==6.3.0
dill==0.3.6
distlib==0.3.6
dlib==19.24.1
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.3.2
exceptiongroup==1.1.0
executing==1.2.0
face-alignment==1.3.5
fastapi==0.88.0
ffmpeg-python==0.2.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.1
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.53.0
h11==0.14.0
hmmlearn==0.2.8
huggingface-hub==0.13.3
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
HyperPyYAML==1.2.0
identify==2.5.18
idna==3.4
imageio==2.28.1
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
inquirer==3.1.3
intervaltree==3.1.0
ipykernel==6.22.0
ipympl==0.9.3
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
julius==0.2.7
jupyter-client==8.1.0
jupyter-core==5.3.0
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
lazy-loader==0.2
lazy-object-proxy==1.9.0
librosa==0.10.0.post2
lightning==2.0.0
lightning-cloud==0.5.32
lightning-utilities==0.7.1
# lilcom==1.6
lit==16.0.0
llvmlite==0.39.1
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
nest-asyncio==1.5.6
networkx==2.8.8
nodeenv==1.7.0
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
optuna==2.10.1
ordered-set==4.1.0
packaging==23.0
pandas==2.0.0
parso==0.8.3
pathtools==0.1.2
pbr==5.11.1
pesq==0.0.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
pluggy==1.0.0
pooch==1.6.0
portalocker==2.7.0
pre-commit==3.1.1
prettytable==3.6.0
primePy==1.3
prompt-toolkit==3.0.38
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
pyDeprecate==0.3.2
Pygments==2.14.0
PyJWT==2.6.0
pylint==2.17.1
pyparsing==3.0.9
pyperclip==1.8.2
pyrootutils==1.0.4
pystoi==0.3.3
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-speech-features==0.6
pytorch-lightning==2.0.1
pytorch-metric-learning==1.7.3
pytz==2023.2
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
readchar==4.0.5
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sacrebleu==2.3.1
# scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
semver==2.13.0
sentencepiece==0.1.97
sentry-sdk==1.18.0
setproctitle==1.3.2
shellingham==1.5.0.post1
simplejson==3.18.4
singledispatchmethod==1.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4
soxr==0.3.4
speechbrain==0.5.14
SQLAlchemy==2.0.4
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
stevedore==5.0.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tokenizers==0.13.2
tomli==2.0.1
tomlkit==0.11.7
toolz==0.12.0
torch==2.0.0
torch-audiomentations==0.11.0
torch-ema==0.3
torch-pitch-shift==1.2.3
torchaudio==2.0.1
torchdata==0.6.0
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.27.4
triton==2.0.0
typer==0.7.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.14
uvicorn==0.21.1
virtualenv==20.19.0
wandb==0.14.0
wcwidth==0.2.6
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
whisper-openai==1.0.0
widgetsnbextension==4.0.7
wrapt==1.15.0
yarl==1.8.2
zipp==3.15.0
